# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.PaginatedUsersResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/user.UserResponse'
        type: array
      pagination:
        $ref: '#/definitions/internal_feature_user_handler.Pagination'
    type: object
...
  internal_feature_user_handler.Pagination:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      totalPages:
        type: integer
    type: object
  internal_feature_user_handler.Response:
    properties:
      data: {}
      status:
        type: integer
      title:
        type: string
      type:
        type: string
    type: object
...
  user.ChangePasswordRequest:
    properties:
      new_password:
        type: string
      old_password:
        type: string
    type: object
  user.CreateUserRequest:
    properties:
      email:
        type: string
      full_name:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  user.UpdateUserRequest:
    properties:
      email:
        type: string
      full_name:
        type: string
      username:
        type: string
    type: object
  user.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      is_disabled:
        type: boolean
      is_email_verified:
        type: boolean
      last_login_at:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: A Golang starter project with Vertical Slice Architecture
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Golang Starter API
  version: "1.0"
paths:
...
  /users:
    get:
      consumes:
      - application/json
      description: Get a paginated list of users
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Sort field and direction (e.g., created_at:desc)
        in: query
        name: sort
        type: string
      - description: Filter expression (e.g., status:active)
        in: query
        name: filter
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_feature_user_handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.PaginatedUsersResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_feature_user_handler.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_feature_user_handler.Response'
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new user
      parameters:
      - description: User details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/internal_feature_user_handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/user.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_feature_user_handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_feature_user_handler.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_feature_user_handler.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_feature_user_handler.Response'
      security:
      - BearerAuth: []
      summary: Create user
      tags:
      - users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_feature_user_handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_feature_user_handler.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_feature_user_handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_feature_user_handler.Response'
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Get a user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_feature_user_handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/user.UserResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_feature_user_handler.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_feature_user_handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_feature_user_handler.Response'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_feature_user_handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/user.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_feature_user_handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_feature_user_handler.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_feature_user_handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_feature_user_handler.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_feature_user_handler.Response'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - users
  /users/{id}/disable:
    post:
      consumes:
      - application/json
      description: Disable a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_feature_user_handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_feature_user_handler.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_feature_user_handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_feature_user_handler.Response'
      security:
      - BearerAuth: []
      summary: Disable user
      tags:
      - users
  /users/{id}/enable:
    post:
      consumes:
      - application/json
      description: Enable (restore) a disabled user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_feature_user_handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_feature_user_handler.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_feature_user_handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_feature_user_handler.Response'
      security:
      - BearerAuth: []
      summary: Enable user
      tags:
      - users
...
  /users/me:
    delete:
      consumes:
      - application/json
      description: Delete the currently authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_feature_user_handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_feature_user_handler.Response'
      security:
      - BearerAuth: []
      summary: Delete current user
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Get the currently authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_feature_user_handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/user.UserResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_feature_user_handler.Response'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update the currently authenticated user
      parameters:
      - description: User details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_feature_user_handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/user.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_feature_user_handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_feature_user_handler.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_feature_user_handler.Response'
      security:
      - BearerAuth: []
      summary: Update current user
      tags:
      - users
  /users/me/change-password:
    post:
      consumes:
      - application/json
      description: Change the current user's password
      parameters:
      - description: Password details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_feature_user_handler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_feature_user_handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_feature_user_handler.Response'
      security:
      - BearerAuth: []
      summary: Change password
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"